+++ src/pyforest/__init__.py
@@ -1,18 +1,6 @@
 # -*- coding: utf-8 -*-
-def get_user_symbols():
-    import inspect
-
-    for index, item in enumerate(inspect.stack()):
-        try:
-            name = item[0].f_globals["__name__"]
-            if name == "__main__":
-                return item[0].f_globals
-        except:  # __name__ attribute does not exist
-            pass
-    return {}
-
-
 from ._imports import *
+from .utils import get_user_symbols
 
 user_symbols = get_user_symbols()
 pyforest_imports = globals().copy().keys()
